pyproject.toml
[project]
name = "tomcat-log-monitor"
version = "1.0.0"
description = "Tomcat WAS 로그 실시간 감시 및 AI 분석 시스템"
authors = [
    { name = "Developer", email = "dev@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
    "streamlit>=1.28.0,<2.0.0",        # Streamlit 안정 버전
    "plotly==5.24.1",                  # 안정화 버전으로 고정
    "pandas>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "langchain>=0.1.0,<0.3.0",
    "langchain-openai>=0.1.0,<0.2.0",
    "openai>=1.0.0,<2.0.0",
    "tqdm>=4.65.0,<5.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0,<8.0.0",
    "black>=23.0.0,<24.0.0",
    "flake8>=6.0.0,<7.0.0"
]